--- src/TypeSelector.module.css
@@ -1,28 +1,0 @@
-.label:has(.radio:active) {
-  filter: var(--filter-darken);
-}
-
-.radio {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 1rem;
-  height: 1rem;
-  flex: none;
-  background: var(--type-color);
-  border: 1px solid var(--color-vibrant-border1);
-  border-radius: 9999px;
-}
-
-.radio:checked {
-  --_bg: black;
-  --_fg: white;
-  background: var(--_fg);
-  box-shadow: inset 0 0 0 4px var(--_bg);
-  border-color: var(--_fg);
-  background-clip: padding-box;
-}
-
-.radio:disabled {
-  background: var(--color-bg2);
-  border-color: var(--color-fg4);
-}
