+++ src/components/Divider.module.css
@@ -1,0 +1,3 @@
+.Divider {
+  border-top: 1px solid var(--color-border3);
+}
