+++ src/components/Divider.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import styles from "./Divider.module.css";
+
+export function Divider() {
+  return <div className={styles.Divider} />;
+}
