+++ src/components/TranslationCard.module.css
@@ -1,0 +1,14 @@
+.TranslationCard_root {
+  color: var(--color-fg1);
+  background-color: var(--color-bg1);
+  border: 1px solid var(--color-border2);
+  border-radius: var(--border-radius3);
+  box-shadow: var(--shadow-button);
+  padding: var(--padding3);
+}
+
+.TranslationCard_pill {
+  padding: 0 var(--padding2);
+  border-radius: var(--border-radius-pill);
+  border: 1px solid var(--color-border3);
+}
