+++ src/components/TypeSelector.module.css
@@ -1,0 +1,64 @@
+.label {
+  padding: var(--padding2) var(--padding3);
+  border: 1px solid;
+  border-radius: var(--border-radius-pill);
+  text-transform: capitalize;
+}
+
+.label:where(:has(.radio:active)) {
+  filter: var(--filter-darken);
+}
+
+.label:where(:has(.radio:checked)) {
+  box-shadow: var(--shadow-button);
+  border-color: var(--color-vibrant-border1);
+  background: var(--type-color-bg);
+  color: white;
+  text-shadow: var(--type-text-shadow);
+}
+
+.label:where(:not(:has(.radio:checked))) {
+  border-color: var(--color-border1);
+  background: var(--background-button-bg);
+  color: var(--color-fg1);
+  box-shadow: var(--shadow-button);
+}
+
+.label:where(:not(:has(.radio:checked))):hover {
+  background: var(--background-button-bg-hover);
+}
+
+/* stylelint-disable-next-line no-descending-specificity */
+.label:where(:has(.radio:disabled)) {
+  pointer-events: none;
+  background: transparent;
+  color: var(--color-fg4);
+  border-color: var(--color-border3);
+  text-shadow: none;
+}
+
+.radio {
+  -webkit-appearance: none;
+  appearance: none;
+  width: 1rem;
+  height: 1rem;
+  flex: none;
+  background: var(--type-color);
+  border: 1px solid var(--color-vibrant-border1);
+  border-radius: 9999px;
+}
+
+.radio:checked {
+  --_bg: black;
+  --_fg: white;
+  background: var(--_fg);
+  box-shadow: inset 0 0 0 4px var(--_bg);
+  border-color: var(--_fg);
+  background-clip: padding-box;
+}
+
+.radio:disabled {
+  background: var(--color-bg2);
+  border-color: var(--color-border3);
+  box-shadow: none;
+}
